+++ src/CommonOPF.jl
@@ -2,8 +2,8 @@ module CommonOPF
 
 using LinearAlgebra
 using JuMP
-using Graphs, MetaGraphs
-import MetaGraphs: inneighbors, outneighbors, induced_subgraph
+using Graphs, MetaGraphsNext
+import MetaGraphsNext: inneighbors, outneighbors, induced_subgraph
 import PowerModelsDistribution: parse_dss, DELTA
 import Logging: SimpleLogger, Error, with_logger